+++ CHANGES.md
@@ -48,6 +48,7 @@ Changes in 5.0
 - Add Free Text to Report Cards form in ReportCards.fnc.php & ReportCards.php
 - Save Template even if no default template found in Template.fnc.php
 - Fix $staff_id for admin user in \_makeLetterGrade.fnc.php
+- Add Min. and Max. Grades option in ReportCards.fnc.php
 
 Changes in 4.9.6
 ----------------